Fixing bugs in getting req.user.id in authMiddleware.js

+++ backend/controllers/goalController.js
@@ -3,24 +3,31 @@ const asyncHandler = require("express-async-handler");
 const db = require("../models");
 
 const Goal = db.goals;
+const User = db.users;
 
+// Problem as of 8/3/22 1047pm req.user.id is undefined. we cannot decode the req.user in auth Middleware
+// Get all goals
 const getAllGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.findAll();
-  res.json(goals);
+  const goals = await Goal.findAll({ where: { user: req.user.id } });
+
+  res.status(200).json(goals);
 });
 
+// Get One Goal
 const getOneGoal = asyncHandler(async (req, res) => {
   const id = req.params.id;
-  const goal = await Goal.findOne({ where: { id: id } });
+
+  const goal = await Goal.findOne({ where: { id: id, user: req.user.id } });
 
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found!");
   }
 
-  res.json(goal);
+  res.status(200).json(goal);
 });
 
+// Post Goal
 const setGoal = asyncHandler(async (req, res) => {
   if (!req.body.text) {
     res.status(400);
@@ -29,7 +36,7 @@ const setGoal = asyncHandler(async (req, res) => {
 
   const data = {
     text: req.body.text,
-    // user: req.user.id,
+    user: req.user.id,
   };
 
   const goal = await Goal.create(data);
@@ -37,6 +44,7 @@ const setGoal = asyncHandler(async (req, res) => {
   res.status(200).json(goal);
 });
 
+// Update Goal
 const updateGoal = asyncHandler(async (req, res) => {
   const goal = await Goal.findByPk(req.params.id);
 
@@ -46,19 +54,19 @@ const updateGoal = asyncHandler(async (req, res) => {
   }
 
   // Make sure User 1 will not be able to UPDATE User 2 goals
-  // const user = await User.findByPk(req.user.id);
+  const user = await User.findByPk(req.user.id);
 
   // Check for user
-  // if (!user) {
-  //   res.status(401);
-  //   throw new Error("User not Found!");
-  // }
+  if (!user) {
+    res.status(401);
+    throw new Error("User not Found!");
+  }
 
   // Make sure the logged in user matches the goal user
-  // if (goal.user !== user.id) {
-  //   res.status(401);
-  //   throw new Error("User not Authorized!");
-  // }
+  if (goal.user !== user.id) {
+    res.status(401);
+    throw new Error("User not Authorized!");
+  }
 
   await Goal.update(req.body, {
     where: { id: req.params.id },
@@ -67,6 +75,7 @@ const updateGoal = asyncHandler(async (req, res) => {
   res.status(200).json("Goal is Updated Successfully!");
 });
 
+// Delete Goal
 const deleteGoal = asyncHandler(async (req, res) => {
   const goal = await Goal.findByPk(req.params.id);
 
@@ -76,19 +85,19 @@ const deleteGoal = asyncHandler(async (req, res) => {
   }
 
   // Make sure User 1 will not be able to DELETE User 2 goals
-  // const user = await User.findByPk(req.user.id);
+  const user = await User.findByPk(req.user.id);
 
   // Check for user
-  // if (!user) {
-  //   res.status(401);
-  //   throw new Error("User not Found!");
-  // }
+  if (!user) {
+    res.status(401);
+    throw new Error("User not Found!");
+  }
 
   // Make sure the logged in user matches the goal user
-  // if (goal.user !== user.id) {
-  //   res.status(401);
-  //   throw new Error("User not Authorized!");
-  // }
+  if (goal.user !== user.id) {
+    res.status(401);
+    throw new Error("User not Authorized!");
+  }
 
   await Goal.destroy({ where: { id: req.params.id } });
   res.status(200).json({ message: "Goal is Deleted Successfully!" });
